fix: fix mentors PR changes request
